WhiteNoise
$ pip install whitenoise

If collectstatic failed during a build, a traceback was provided that will be helpful in diagnosing the problem. If you need additional information about the environment collectstatic was run in, use the DEBUG_COLLECTSTATIC configuration.

$ heroku config: set DEBUG_COLLECTSTATIC = 1
Sometimes, you may not want Heroku to run collectstatic on your behalf. You can disable the collectstatic build step with the DISABLE_COLLECTSTATIC configuration:

$ heroku config: set DISABLE_COLLECTSTATIC = 1